Move unit tests into the tests directory
