+++ src/net/mod.rs
@@ -28,7 +28,7 @@
 //! # }) }
 //! ```
 
-pub use self::addr::{SocketAddr, SocketAddrV4, SocketAddrV6, ToSocketAddrs};
+pub use addr::{SocketAddr, SocketAddrV4, SocketAddrV6, ToSocketAddrs};
 pub use tcp::{Incoming, TcpListener, TcpStream};
 pub use udp::UdpSocket;
